requirements.txt
geopy
requests
google-cloud-bigquery
pycountry
pandas
openpyxl
pyarrow